refactor
refactored esther.pyinto test_us12.py and test_us7.py and added marriage dates to all families
