+++ src/inputB.ged
@@ -651,6 +651,9 @@
 1 CHIL @I3@
 1 CHIL @I4@
 1 CHIL @I5@
+1 MARR
+2 DATE 8 NOV 1995
+2 PLAC Monroe, NC
 1 _CURRENT Y
 0 @F2@ FAM
 1 HUSB @I7@
@@ -678,6 +681,9 @@
 1 CHIL @I26@
 1 CHIL @I27@
 1 _CURRENT Y
+1 MARR
+2 DATE 1 DEC 1939
+2 PLAC Logan, UT
 0 @F5@ FAM
 1 HUSB @I7@
 1 WIFE @I12@
@@ -694,34 +700,55 @@
 1 HUSB @I14@
 1 WIFE @I13@
 1 CHIL @I7@
+1 MARR
+2 DATE 1 DEC 1930
+2 PLAC Logan, UT
 1 _CURRENT N
 0 @F7@ FAM
 1 HUSB @I19@
 1 WIFE @I18@
 1 CHIL @I13@
+1 MARR
+2 DATE 1 DEC 1909
+2 PLAC Logan, UT
 1 _CURRENT Y
 0 @F8@ FAM
 1 HUSB @I14@
 1 WIFE @I15@
+1 MARR
+2 DATE 1 DEC 1970
+2 PLAC Logan, UT
 1 _CURRENT Y
 0 @F9@ FAM
 1 HUSB @I16@
 1 WIFE @I17@
 1 CHIL @I14@
+1 MARR
+2 DATE 1 SEP 1905
+2 PLAC Pueblo, CO
 1 _CURRENT Y
 0 @F10@ FAM
 1 HUSB @I30@
 1 WIFE @I22@
 1 CHIL @I32@
+1 MARR
+2 DATE 1 DEC 1960
+2 PLAC Houston, TX
 1 _CURRENT N
 0 @F14@ FAM
 1 HUSB @I30@
 1 WIFE @I31@
+1 MARR
+2 DATE 1 DEC 1930
+2 PLAC Logan, UT
 1 _CURRENT Y
 0 @F11@ FAM
 1 HUSB @I29@
 1 WIFE @I28@
 1 CHIL @I22@
+1 MARR
+2 DATE 1 DEC 1884
+2 PLAC Houston, TX
 1 _CURRENT Y
 0 @F12@ FAM
 1 HUSB @I25@
@@ -737,19 +764,31 @@
 1 CHIL @I44@
 1 CHIL @I45@
 1 CHIL @I47@
+1 MARR
+2 DATE 1 DEC 1887
+2 PLAC Logan, UT
 1 _CURRENT Y
 0 @F13@ FAM
 1 HUSB @I34@
 1 WIFE @I33@
 1 CHIL @I25@
+1 MARR
+2 DATE 1 DEC 1861
+2 PLAC Logan, UT
 1 _CURRENT Y
 0 @F15@ FAM
 1 HUSB @I38@
 1 WIFE @I37@
+1 MARR
+2 DATE 1 DEC 1935
+2 PLAC Logan, UT
 1 _CURRENT Y
 0 @F16@ FAM
 1 HUSB @I42@
 1 WIFE @I41@
+1 MARR
+2 DATE 1 DEC 1869
+2 PLAC Logan, UT
 1 _CURRENT Y
 0 @F17@ FAM
 1 HUSB @I52@
@@ -758,5 +797,8 @@
 1 CHIL @I49@
 1 CHIL @I50@
 1 CHIL @I51@
+1 MARR
+2 DATE 1 DEC 1869
+2 PLAC Logan, UT
 1 _CURRENT N
 0 TRLR
